set protection for admin routes

+++ backend-laravel/routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Artisan;
 use App\Http\Controllers\API\AuthController;
@@ -31,27 +32,46 @@ Route::middleware('api')->group(function () {
         }
     });
 
-    Route::get('/storage-link-debug', function () {
-        try {
-            Artisan::call('storage:link');
-            return nl2br(Artisan::output());
-        } catch (\Exception $e) {
-            return $e->getMessage();
-        }
-    });
-
-
-    Route::get('/reset', function () {
-        Artisan::call('migrate:reset', ['--force' => true]);
-        Artisan::call('migrate', ['--force' => true]);
-        return 'Migration reset and re-run!';
-    });
-
-    Route::get('/check-db', function () {
-        return DB::select('SELECT DATABASE() AS db_name');
-    });
-
     Route::middleware('jwt.auth')->group(function () {
+
+        Route::get('/storage-link-debug', function () {
+            try {
+                if (Auth::user() && Auth::user()->role == ROLE_ADMIN) {
+                    Artisan::call('storage:link');
+                    return nl2br(Artisan::output());
+                } else {
+                    abort(403, 'Unauthorized: Admins only');
+                }
+            } catch (\Exception $e) {
+                return $e->getMessage();
+            }
+        });
+
+        Route::get('/reset', function () {
+            if (Auth::user() && Auth::user()->role == ROLE_ADMIN) {
+                try {
+                    Artisan::call('migrate:reset', ['--force' => true]);
+                    Artisan::call('migrate', ['--force' => true]);
+                    return nl2br("Migration reset and re-run!\n\n" . Artisan::output());
+                } catch (\Exception $e) {
+                    return response($e->getMessage(), 500);
+                }
+            }
+            abort(403, 'Unauthorized: Admins only');
+        });
+
+        Route::get('/check-db', function () {
+            if (Auth::user() && Auth::user()->role == ROLE_ADMIN) {
+                try {
+                    $db = DB::select('SELECT DATABASE() AS db_name');
+                    return response()->json($db);
+                } catch (\Exception $e) {
+                    return response($e->getMessage(), 500);
+                }
+            }
+            abort(403, 'Unauthorized: Admins only');
+        });
+
         Route::get('/user',   [AuthController::class, 'me']);
         Route::post('/logout', [AuthController::class, 'logout']);
